Fix time conversion bug in Datetime code
The Datetime code had a couple bugs in the datetime -> seconds-since-epoch conversion code that caused it to be off by a day or two in leap years.
These were mostly because when looking at whether to account for a leap year, they were evaluating the previous year/month rather than the current year/month.
This changes the logic to correctly account for leap years.