Make sure we're not using the PLL as main clock when resetting it. (#424)
This PR resets the main clock source to FFRO divided by 4 before
changing the PLL configuration.

When a bootloader and payload application both use `embassy-imxrt`, the
clock configuration routine is ran twice. The second time, the PLL will
be reset while it's also being used as a main clock source. The result
is that the CPU stops executing instructions and even the debug probe
gets disconnected (which made this a bit difficult to debug
:sweat_smile:).

By resetting the main clock source to the reset default, we can safely
re-configure the PLL, and then switch (back) to it.